Refactor: Modularize API client and services
Split the monolithic `api.ts` file into smaller, domain-specific modules (`auth.ts`, `client.ts`, `employees.ts`, `services.ts`). This improves code organization, maintainability, and readability by separating concerns. Updated all affected components and pages to use the new import paths.
