--- frontend/src/api/api.ts
@@ -1,145 +1,0 @@
-import axios from 'axios';
-
-const apiClient = axios.create({
-  baseURL: 'http://127.0.0.1:8000/api',
-});
-
-apiClient.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem('accessToken');
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
-
-apiClient.interceptors.response.use(
-  (response) => {
-    return response;
-  },
-  async (error) => {
-    const originalRequest = error.config;
-
-    if (error.response.status === 401 && !originalRequest._retry) {
-      originalRequest._retry = true; // Marca como uma tentativa de repetição
-
-      try {
-        // Tenta obter um novo token usando o refresh token
-        const refreshToken = localStorage.getItem('refreshToken');
-        const response = await axios.post('http://127.0.0.1:8000/api/token/refresh/', {
-          refresh: refreshToken,
-        });
-
-        const newAccessToken = response.data.access;
-        localStorage.setItem('accessToken', newAccessToken);
-        apiClient.defaults.headers.common['Authorization'] = `Bearer ${newAccessToken}`;
-
-        originalRequest.headers['Authorization'] = `Bearer ${newAccessToken}`;
-
-        return apiClient(originalRequest);
-
-      } catch (refreshError) {
-        console.error("Refresh token inválido ou expirado.", refreshError);
-        localStorage.removeItem('accessToken');
-        localStorage.removeItem('refreshToken');
-        window.location.href = '/login';
-        return Promise.reject(refreshError);
-      }
-    }
-    return Promise.reject(error);
-  }
-);
-
-export interface CreateEmployee {
-  username: string;
-  first_name: string;
-  last_name: string;
-  email: string;
-  password: string;
-  phone?: string;
-}
-
-export interface UpdateEmployee {
-  first_name?: string;
-  last_name?: string;
-  email?: string;
-  password?: string;
-  username?: string;
-  phone?: string;
-}
-
-export interface Employee {
-  id: number;
-  username: string;
-  first_name: string;
-  last_name: string;
-  email: string;
-  phone?: string;
-  role: string;
-  is_active: boolean;
-}
-
-export const getEmployees = async (): Promise<Employee[]> => {
-  try {
-    const response = await apiClient.get('/employees/');
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching employees:", error);
-    throw error;
-  }
-};
-
-export const getEmployee = async (id: number): Promise<Employee> => {
-  try {
-    const response = await apiClient.get(`/employees/${id}/`);
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching employee:", error);
-    throw error;
-  }
-};
-
-export const updateEmployee = async (id: number, employeeData: UpdateEmployee): Promise<Employee> => {
-  try {
-    const response = await apiClient.patch(`/employees/${id}/`, employeeData);
-    return response.data;
-  } catch (error) {
-    console.error("Error updating employee:", error);
-    throw error;
-  }
-};
-
-export const createEmployee = async (employeeData: CreateEmployee) => {
-  try {
-    const response = await apiClient.post('/employees/', employeeData);
-    return response.data;
-  } catch (error) {
-    console.error("Error creating employee:", error);
-    throw error;
-  }
-};
-
-export const deleteEmployee = async (id: number): Promise<void> => {
-  try {
-    await apiClient.delete(`/employees/${id}/`);
-  } catch (error) {
-    console.error("Error deleting employee:", error);
-    throw error;
-  }
-};
-
-export const getMe = async (): Promise<Employee> => {
-  try {
-    const response = await apiClient.get('/me/');
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching current user:", error);
-    throw error;
-  }
-};
-
-export default apiClient;
