+++ frontend/src/api/index.ts
@@ -1,0 +1,4 @@
+export * from './auth';
+export * from './employees';
+export * from './services';
+//export * from './appointments';
